feat(epic-e/drainer+db): wire CustomerScoreService into order + customer handlers
- apps/drainer/src/handlers/order.ts: after orderRepo.upsertFromWebhook,
  call CustomerScoreService.recompute inline in the SAME tx when
  customerId !== null. Guest checkouts (customerId null) skip scoring per
  P-10. Discriminated log captures skip reason, branchTaken, state delta,
  cohort size, durationMs (S-10 observability hook).
- apps/drainer/src/handlers/customer.ts: runUpsert (create + updated)
  calls recompute after the customer upsert in the SAME tx.
  handleCustomerDeleted does NOT call recompute per §S-7 (soft delete
  excludes from future cohort reads; existing CustomerScore row stays
  forensic).
- apps/drainer/src/dispatch.ts: comment update on customer.state_changed —
  producer ships this session, consumer is Epic G; routing stays at
  handleNoop so the future swap is a single case-body replacement.
- packages/db/src/repositories/order.repository.ts: added customerId to
  UpsertOrderFromWebhookResult. Additive (existing tests use property-
  level assertions); needed so the drainer can forward it to the scoring
  call without a second customer lookup.

Two real bugs surfaced by integration testing + fixed in this batch:
- packages/db/src/tenant-scope.ts: added 'CustomerScore' to
  TENANT_SCOPED_MODELS. Outright gap from batch 1 — without it the Prisma
  extension passed CustomerScore reads + writes through with no tenant
  assertion or injection. Tenant safety bug independent of scoring
  correctness.
- packages/db/src/repositories/customer-score.repository.ts: SUM cast.
  Postgres widens SUM(bigint) to numeric (overflow protection); Prisma
  deserializes numeric as Decimal; tx.customerScore.upsert({mCents: Decimal})
  throws "Expected BigInt, provided Float". Explicit ::bigint cast in
  raw SQL keeps the column-native type end-to-end. JSDoc paragraph
  explains the rationale (cents max ~9.2e18 = $92 quintillion — safe
  margin against overflow). Would have caused every production paid-
  order recompute to fail without integration testing surfacing it.

Tests:
- apps/drainer/tests/handlers/order.test.ts: 10 → 15 (+5). New mock
  surface for CustomerScoreService + the two repo deps. New describe
  "scoring service integration": customerId-null SKIPPED (guest
  checkout); scoring throw propagates; customer_not_found skip benign;
  state-change result flows through; recompute receives same tx as
  upsert.
- apps/drainer/tests/handlers/customer.test.ts: 8 → 12 (+4). Scoring
  invoked on create/updated paths; explicit negative test that
  handleCustomerDeleted NEVER calls recompute (§S-7 lock); scoring throw
  propagates.
- Full workspace: 627 → 636 unit tests. Drainer int 20/20 (new wiring
  exercised end-to-end). Web int 27/27.

Doc:
- EPIC-E-SESSION-2-DESIGN.md: corrected SQL spec in the readCohort
  docstring to include the ::bigint cast on SUM.

Pre-existing flake noted (NOT a regression from this batch):
- packages/db/tests/integration/smoke.test.ts > "T1 functional + partial
  index": planner-state-dependent; with 1 customer row Postgres picks
  Seq Scan over Index Scan. Test was last touched 2026-Q1 (5 commits
  pre-Epic-E); none of my batch-4 changes affect Customer or its
  indexes. Fix is a separate test-fixture or SET enable_seqscan = off.

+++ packages/db/src/repositories/customer-score.repository.ts
@@ -51,10 +51,19 @@ export class CustomerScoreRepository extends BaseRepository {
    * quintile math, and makes the method deterministic under unit test.
    *
    * Result row types:
-   *   - `customerId` TEXT  → `string`
-   *   - `rDays`      INT   → `number`
-   *   - `fCount`     INT   → `number`
-   *   - `mCents`     BIGINT → `bigint`  (Prisma 5 maps BIGINT to BigInt)
+   *   - `customerId` TEXT             → `string`
+   *   - `rDays`      INT (cast in SQL) → `number`
+   *   - `fCount`     INT (cast in SQL) → `number`
+   *   - `mCents`     BIGINT (cast in SQL) → `bigint`
+   *
+   * Why the explicit `::bigint` cast on the SUM: Postgres widens
+   * `SUM(bigint)` to `numeric` to avoid overflow on huge sums. Prisma
+   * deserializes `numeric` as `Decimal` (not BigInt), and feeding a
+   * `Decimal` to `tx.customerScore.upsert({ mCents })` throws
+   * `PrismaClientValidationError: Expected BigInt, provided Float`.
+   * The `::bigint` cast keeps the column-native type end-to-end and
+   * is safe because no realistic single-customer 365d sum approaches
+   * `bigint` max (~9.2 × 10¹⁸ cents = $92 quintillion).
    *
    * No `deletedAt` predicate: `Order` is immutable in this schema
    * (cancel / refund are status changes, not soft deletes) — see
@@ -74,7 +83,7 @@ export class CustomerScoreRepository extends BaseRepository {
       SELECT "customerId",
              FLOOR(EXTRACT(EPOCH FROM (${now}::timestamptz - MAX(COALESCE("shopifyProcessedAt", "placedAt")))) / 86400)::int AS "rDays",
              COUNT(*)::int AS "fCount",
-             SUM("totalAmountCents") AS "mCents"
+             SUM("totalAmountCents")::bigint AS "mCents"
       FROM "Order"
       WHERE "merchantId" = ${merchantId}
         AND "financialStatus" = 'paid'
